Fixing off-by-one error in ethsnarks::bit_list_to_ints()
bit_list_to_ints() populates a word vector from a list of bits, given a word size
desired for each element of the returned vector.  There is an off-by-one error
in its calculation of the number of words that causes it to overwrite the region
of its allocated vector by one.

For example, if you have a 256 bit input and a word size of 8, the routine will
attempt to read populate 33 bytes into its return vector, reading 8 bits beyond the end of its input bit list.

This was causing the test_hashpreimage test case to report a successful test result, but still fault with an 'Abort Trap: 6' error.  It may also have been causing a more direct false negative in the test_sha256_full test case.

I was only analyzing test_hashpreimage when I found and corrected this problem,
but it appears to have also accidentally fixed test_sha256_full, although I have
not done enough analysis to be able to offer a guess as to why.